requirements.txt
numpy
opencv-python
pillow
dlib
mediapipe
deepface
tensorflow 